build.sh: clean ups
Even if the /s option is given, the PO2013 won't write a new symbol
file if an accurate file already exists on disk. This gives a bit of
trouble in this script, as the compiler doesn't generate symbol files
in stage 3. Hence build2 and build3 differ, even if everything went
well.

Fix this by hiding the symbol files from the compiler.