Silence bogus GCC warning about strncpy
GCC 8 warns about possible string truncation. But files_check_name
already verified that the filename fits in a NameLength buffer. GCC is
just not smart enough to see that.

In fact strcpy would be safe here. But strncpy is nice because it
zeroes out the rest of the buffer. PO2013's FileDir.Enumerate does
that as well, I think.  Adding an assert() silences the GCC warning.